change all instances of projects to project
