+++ app.js
@@ -1,20 +1,20 @@
 require("dotenv").config();
 const express = require("express");
 const app = express();
-;(async () => {
+(async () => {
   // Add headers before the routes are defined
-    app.use(require("./middleware/headers"));
-    const controllers = require("./controllers");
-    app.use(express.json());
+  app.use(require("./middleware/headers"));
+  const controllers = require("./controllers");
+  app.use(express.json());
 
-    app.use("/user", controllers.userController);
-    app.use(require("./middleware/validate-session"));
-    app.use("/employee", controllers.employeeController);
-    app.use("/company", controllers.companyController);
-    app.use("/timesheet", controllers.timesheetController);
-    app.use("/project", controllers.projectsController);
+  app.use("/user", controllers.userController);
+  app.use(require("./middleware/validate-session"));
+  app.use("/employee", controllers.employeeController);
+  app.use("/company", controllers.companyController);
+  app.use("/timesheet", controllers.timesheetController);
+  app.use("/project", controllers.projectController);
 
-    app.listen(process.env.PORT, () => {
-        console.log(`[Server]: App is listening at port: ${process.env.PORT}`);
-    });
+  app.listen(process.env.PORT, () => {
+    console.log(`[Server]: App is listening at port: ${process.env.PORT}`);
+  });
 })();
