+++ controllers/index.js
@@ -1,7 +1,7 @@
 module.exports = {
-    userController: require("./usercontroller"),
-    employeeController: require("./employeecontroller"),
-    companyController: require("./companycontroller"),
-    timesheetController: require("./timesheetcontroller"),
-    projectsController: require("./projectscontroller"),
+  userController: require("./usercontroller"),
+  employeeController: require("./employeecontroller"),
+  companyController: require("./companycontroller"),
+  timesheetController: require("./timesheetcontroller"),
+  projectController: require("./projectcontroller"),
 };
